run prettier over permission.ts
